requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
attrs==23.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.0.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.1
django-cache-url==3.4.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
duplicity==0.8.21
environs==11.0.0
evdev==1.6.1
fasteners==0.14.1
filelock==3.6.0
Flask==2.1.0
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
gunicorn==21.2.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
ipython==7.31.1
itsdangerous==2.1.2
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
parso==0.8.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.12.4
psycopg==3.1.18
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.11.1
pycups==2.0.1
PyDrive==1.3.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.7.6
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-xlib==0.33
pythran==0.10.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.33.0
reportlab==3.6.8
requests==2.25.1
rpds-py==0.18.0
rsa==4.8
scapy==2.4.4
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.4
sympy==1.9
systemd-python==234
traitlets==5.1.1
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
xdg==5
xkit==0.0.0
zipp==1.0.0
